Notes and Code for Ch. 3.6 to Chapter 4
This commit updates the notes and code to match the functionality of the
book between Chapter 3.6 through to Chapter 4. This chapter was
primarily focused on adding Conditionals, Bindings, Functions, Function
Calls, Error Handling, and the Environment of our Monkey programming
language. The book also discussed Garbage Collection and how we're going
to use Go's Garbage Collector due to the degree of difficulty of turning
off Go's garbage collection.